* add entt group benchmarks for iterations     * update frameworks and benchmarks with the new BenchmarkSuite     * split benchmarks into more sub projects     * make benchmarks more optional

+++ src/base/base/components/DataComponent.h
@@ -6,8 +6,8 @@
 namespace ecs::benchmarks::base::components {
 
     struct DataComponent {
-        float thingy{0.0F};
-        int dingy{0};
+        int thingy{0};
+        double dingy{0.0F};
         bool mingy{false};
         std::string stringy;
     };
